+++ docs/version-history.md
@@ -7,7 +7,8 @@ Version 0.4 remains in beta testing. The main change from version 0.3 is includi
 **Release date:** forthcoming
 
   * Added `osm` function for osmotic coefficient to **matrix** module;
-  * Added **io** module functions to roughly convert between molinity (mol/kg-solution) and molality (mol/kg-H<sub>2</sub>O).
+  * Added **io** module functions to roughly convert between molinity (mol/kg-solution) and molality (mol/kg-H<sub>2</sub>O);
+  * Added `salinity2mols` function in **io** module to estimate molalities from salinity for [MZF93](../references/#MZF93) artificial seawater.
 
 ## 0.4.2